create full constructor for organisms, human class

+++ include/Organism.h
@@ -19,7 +19,7 @@ protected:
     World* world;
     int32_t lastActionTurn = -1;
 
-    Organism(uint16_t x, uint16_t y, uint8_t power, uint8_t initiative, World* world);
+    Organism(uint16_t x, uint16_t y, uint8_t power, uint16_t initiative, World* world);
     ~Organism();
     static direction getRandomDirection();
     static bool isEveryDirectionChecked(const bool* directions);
